+++ scripts/package_antares_web.sh
@@ -9,7 +9,7 @@
 set -e
 
 ANTARES_SOLVER_VERSION="8.8"
-ANTARES_SOLVER_FULL_VERSION="8.8.4"
+ANTARES_SOLVER_FULL_VERSION="8.8.5"
 ANTARES_SOLVER_VERSION_INT="880"
 
 SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd -P)
@@ -44,11 +44,6 @@ else
   pyinstaller --distpath ${DIST_DIR} AntaresWebLinux.spec
   popd
 fi
-
-echo "INFO: Generating the Worker Application..."
-pushd ${PROJECT_DIR}
-pyinstaller --distpath ${DIST_DIR} AntaresWebWorker.spec
-popd
 
 echo "INFO: Creating destination directory '${ANTARES_SOLVER_DIR}'..."
 mkdir -p "${ANTARES_SOLVER_DIR}"
@@ -91,7 +86,7 @@ if [[ "$OSTYPE" == "msys"* ]]; then
   cp "${RESOURCES_DIR}/AntaresWebServerShortcut.lnk" "${DIST_DIR}"
 else
   echo "INFO: Updating executable permissions..."
-  for excutable in "${DIST_DIR}/AntaresWeb/AntaresWebServer" "${DIST_DIR}/AntaresWeb/AntaresTool" "${DIST_DIR}/AntaresWebWorker"; do
+  for excutable in "${DIST_DIR}/AntaresWeb/AntaresWebServer" "${DIST_DIR}/AntaresWeb/AntaresTool"; do
     chmod +x "${excutable}"
   done
 fi